requirements.txt
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.6.1
requests==2.31.0
SQLAlchemy==2.0.23
python-multipart==0.0.6
PyPDF2==3.0.1
python-pptx==0.6.21
